Fix: reflecting the feedback
