+++ src/pages/heeyeonKim/LeftArticle/CommentInput/CommentInput.scss
@@ -1,0 +1,29 @@
+.comment-form-wrapper {
+  .comment-box {
+    flex: 1;
+    overflow: hidden;
+    border: none;
+    outline: none;
+    resize: none;
+
+    &::placeholder {
+      color: #c2bebe;
+      font-size: 16px;
+    }
+  }
+
+  .btn-post-comment {
+    border: none;
+    background-color: white;
+    color: #c4e1fb;
+    font-size: 16px;
+    cursor: pointer;
+  }
+  .btn-post-comment-blue {
+    border: none;
+    background-color: white;
+    color: #499ff1;
+    font-size: 16px;
+    cursor: pointer;
+  }
+}
