+++ src/pages/heeyeonKim/RightSection/RightSection.scss
@@ -1,0 +1,92 @@
+.div-right {
+  .div-right-wrapper {
+    display: flex;
+    flex-direction: column;
+    width: 360px;
+    gap: 24px;
+  }
+
+  .me-profile {
+    display: flex;
+    gap: 16px;
+    align-items: center;
+  }
+
+  .friends-list-wrapper {
+    border: 1px solid #c2bebe;
+  }
+
+  .text-setting {
+    display: flex;
+    flex-direction: column;
+    padding-right: 52px;
+    line-height: 1.3;
+    gap: 16px;
+  }
+}
+
+.me-profile {
+  &-img {
+    width: 64px;
+    border-radius: 50px;
+  }
+
+  &-info {
+    display: flex;
+    flex-direction: column;
+    gap: 8px;
+  }
+}
+
+.friends-list-wrapper {
+  .friends-list-header {
+    display: flex;
+    justify-content: space-between;
+    padding: 12px 12px 0 12px;
+  }
+
+  .friends-list {
+    display: flex;
+    flex-direction: column;
+    padding: 16px;
+    gap: 16px;
+  }
+
+  .friends-info {
+    display: flex;
+    align-items: center;
+    gap: 8px;
+
+    &.follow-replace {
+      justify-content: space-between;
+    }
+  }
+
+  .friends-img {
+    width: 56px;
+    border-radius: 50px;
+  }
+
+  .friends-main-info {
+    display: flex;
+    align-items: center;
+    gap: 8px;
+  }
+}
+
+.show-time {
+  padding: 4px 12px 8px 12px;
+  color: #c2bebe;
+}
+
+.blue-button {
+  color: #499ff1;
+}
+
+.font-color-gray {
+  color: #c2bebe;
+}
+
+.font-color-blue {
+  color: #499ff1;
+}
